Add files via upload
Bus Reservation System
Using Trubo c++ compiler to run this code. This is a simple bus reservation system programmed using C++. The program allows you to:

Add bus details
Reserve a bus seat according to the vacant seat available
Check for the list of vacant seats in a bus
See the available buses for now
The program is made up of the following components:

Route Class: This class contains information about the available routes.
Bus Class: The bus class is used to store the data about the bus and the number of seats.
Ticket Class: This class stores the information about the ticket, such as the passenger name, seat number, and bus number.
User Class: This class stores the information about the user, such as the username and password.
Bus Reservation System Class: This class is the main class that controls the entire system.
To use the program, you first need to create an instance of the Bus Reservation System class. Then, you can call the various methods on this class to perform the different operations.

For example, to add a new bus, you can call the addBus() method. To reserve a seat, you can call the reserveSeat() method. To check the list of vacant seats, you can call the getVacantSeats() method. And to see the available buses, you can call the getAvailableBuses() method.

The program also provides a simple user interface that allows you to interact with the system. To start the user interface, you can call the startUI() method on the Bus Reservation System class.

This is just a simple example of a bus reservation system in C++. You can extend the program to add more features, such as the ability to cancel tickets, search for buses by route, and so on.